fix: resolve Codacy issues in smart component tests from SITE-5421
Move StubComponent stub into its own file and replace json_encode()
with wp_json_encode() per WordPress coding standards. Issues originated
in commit 3605953, not the PCC-3253 change.
